% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_open_data.R
\docType{data}
\name{base_ckan_url}
\alias{base_ckan_url}
\title{Open Data CKAN API Utils
Base URL for the CKAN API}
\format{
An object of class \code{character} of length 1.
}
\usage{
base_ckan_url
}
\description{
Open Data CKAN API Utils
Base URL for the CKAN API
}
\keyword{datasets}
